Add initial python implementation of GridPlot
There are some things wrong with this, but it gets the basic in place so
we can start playing with it.
