+++ bokeh/models/layouts.py
@@ -8,11 +8,22 @@ import logging
 logger = logging.getLogger(__name__)
 
 from ..core import validation
-from ..core.validation.warnings import EMPTY_LAYOUT, BOTH_CHILD_AND_ROOT
-from ..core.properties import abstract, Bool, Int, Instance, List, Responsive
+
+from ..core.validation.warnings import (
+    EMPTY_LAYOUT,
+    BOTH_CHILD_AND_ROOT,
+    MISSING_RENDERERS,
+    NO_DATA_RENDERERS,
+)
+from ..core.validation.errors import REQUIRED_RANGE
+from ..core.enums import Location
+from ..core.properties import abstract, Bool, Int, Instance, List, Responsive, Enum
 from ..embed import notebook_div
 from ..model import Model
 from ..util.deprecate import deprecated
+
+from ..core.query import find
+from ..util.plot_utils import _list_attr_splat, _select_helper
 
 
 @abstract
@@ -163,6 +174,86 @@ class Column(Box):
     Children can be specified as positional arguments, as a single argument
     that is a sequence, or using the ``children`` keyword argument.
     """
+
+
+class GridPlot(Box):
+    """ A 2D grid of plots rendered on separate canvases.
+
+    """
+    def __init__(self, *args, **kwargs):
+        if len(args) > 0 and "children" in kwargs:
+            raise ValueError("'children' keyword cannot be used with positional arguments")
+        elif len(args) > 0:
+            nested_children = list(args)
+        else:
+            nested_children = kwargs['children']
+
+        toolbar_location = kwargs.get('toolbar_location', 'left')
+        responsive = kwargs.get('responsive', 'box')
+
+        # Make the grid
+        tools = []
+        rows = []
+
+        for row in nested_children:
+            row_tools = []
+            for plot in row:
+                if plot:
+                    row_tools = row_tools + plot.toolbar.tools
+                    plot.toolbar_location = None
+                    plot.responsive = responsive
+            tools = tools + row_tools
+            rows.append(Row(children=row, responsive=responsive))
+
+        grid = Column(children=rows, responsive=responsive)
+
+        # Make the toolbar
+        orientation = 'vertical'
+        if toolbar_location in ['above', 'below']:
+            orientation = 'horizontal'
+
+        from bokeh.models.tools import ToolbarBox
+        toolbar = ToolbarBox(
+            tools=tools,
+            orientation=orientation,
+            merge_tools=True,
+            responsive=responsive
+        )
+
+        # Assign the new children
+        if toolbar_location in ['left', 'above']:
+            kwargs['children'] = [toolbar, grid]
+        else:
+            kwargs['children'] = [grid, toolbar]
+
+        super(GridPlot, self).__init__(*args, **kwargs)
+
+    @validation.error(REQUIRED_RANGE)
+    def _check_required_range(self):
+        pass
+
+    @validation.warning(MISSING_RENDERERS)
+    def _check_missing_renderers(self):
+        pass
+
+    @validation.warning(NO_DATA_RENDERERS)
+    def _check_no_data_renderers(self):
+        pass
+
+    @validation.warning(EMPTY_LAYOUT)
+    def _check_empty_layout(self):
+        from itertools import chain
+        if not list(chain(self.children)):
+            return str(self)
+
+    border_space = Int(0, help="""
+    Distance (in pixels) between adjacent plots.
+    """)
+
+    toolbar_location = Enum(Location, default="left", help="""
+    Where the toolbar will be located. If set to None, no toolbar
+    will be attached to the plot.
+    """)
 
 
 # ---- DEPRECATIONS
