+++ bokeh/models/plots.py
@@ -30,47 +30,7 @@ from .sources import DataSource, ColumnDataSource
 from .tools import Tool, ToolEvents, Toolbar
 from .layouts import LayoutDOM
 
-
-class _list_attr_splat(list):
-    def __setattr__(self, attr, value):
-        for x in self:
-            setattr(x, attr, value)
-
-    def __dir__(self):
-        if len(set(type(x) for x in self)) == 1:
-            return dir(self[0])
-        else:
-            return dir(self)
-
-def _select_helper(args, kwargs):
-    """
-    Allow flexible selector syntax.
-    Returns:
-        a dict
-    """
-    if len(args) > 1:
-        raise TypeError("select accepts at most ONE positional argument.")
-
-    if len(args) > 0 and len(kwargs) > 0:
-        raise TypeError("select accepts EITHER a positional argument, OR keyword arguments (not both).")
-
-    if len(args) == 0 and len(kwargs) == 0:
-        raise TypeError("select requires EITHER a positional argument, OR keyword arguments.")
-
-    if args:
-        arg = args[0]
-        if isinstance(arg, dict):
-            selector = arg
-        elif isinstance(arg, string_types):
-            selector = dict(name=arg)
-        elif issubclass(arg, Model):
-            selector = {"type": arg}
-        else:
-            raise RuntimeError("Selector must be a dictionary, string or plot object.")
-
-    else:
-        selector = kwargs
-    return selector
+from ..util.plot_utils import _list_attr_splat, _select_helper
 
 
 class Plot(LayoutDOM):
@@ -693,87 +653,3 @@ class Plot(LayoutDOM):
     """)
 
     responsive = Override(default='fixed')
-
-
-class GridPlot(LayoutDOM):
-    """ A 2D grid of plots rendered on separate canvases.
-
-    """
-
-    @validation.error(REQUIRED_RANGE)
-    def _check_required_range(self):
-        pass
-
-    @validation.warning(MISSING_RENDERERS)
-    def _check_missing_renderers(self):
-        pass
-
-    @validation.warning(NO_DATA_RENDERERS)
-    def _check_no_data_renderers(self):
-        pass
-
-    @validation.warning(EMPTY_LAYOUT)
-    def _check_empty_layout(self):
-        from itertools import chain
-        if not list(chain(self.children)):
-            return str(self)
-
-    children = List(List(Instance(Plot)), default=[[]], help="""
-    An array of plots to display in a grid, given as a list of lists of
-    Plot objects. To leave a position in the grid empty, pass None for
-    that position in the ``children`` list.
-    """)
-
-    border_space = Int(0, help="""
-    Distance (in pixels) between adjacent plots.
-    """)
-
-    toolbar_location = Enum(Location, default="left", help="""
-    Where the toolbar will be located. If set to None, no toolbar
-    will be attached to the plot.
-    """)
-
-    responsive = Override(default='fixed')
-
-    def select(self, *args, **kwargs):
-        ''' Query this object and all of its references for objects that
-        match the given selector. See Plot.select for detailed usage information.
-
-        Returns:
-            seq[Model]
-        '''
-
-        selector = _select_helper(args, kwargs)
-
-        # Want to pass selector that is a dictionary
-        return _list_attr_splat(find(self.references(), selector, {'gridplot': self}))
-
-    def column(self, col):
-        ''' Return a given column of plots from this GridPlot.
-
-        Args:
-            col (int) : index of the column to return
-
-        Returns:
-            seq[Plot] : column of plots
-
-        '''
-        try:
-            return [row[col] for row in self.children]
-        except:
-            return []
-
-    def row(self, row):
-        ''' Return a given row of plots from this GridPlot.
-
-        Args:
-            rwo (int) : index of the row to return
-
-        Returns:
-            seq[Plot] : row of plots
-
-        '''
-        try:
-            return self.children[row]
-        except:
-            return []
